Create a Docs specific ExampleRunner
Contains all docs related side effects eg, getting the example url and files
