+++ packages/ag-charts-website/src/constants.ts
@@ -51,5 +51,5 @@ export const SITE_BASE_URL_SEGMENTS = SITE_BASE_URL.split('/').filter(Boolean).l
  */
 export const DEV_FILE_BASE_PATH = '/dev';
 
-export const FRAMEWORK_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 1;
-export const PAGE_NAME_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 2;
+export const DOCS_FRAMEWORK_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 1;
+export const DOCS_PAGE_NAME_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 2;
