+++ packages/ag-charts-website/src/pages/[internalFramework]/[pageName]/examples/[exampleName].astro
@@ -2,7 +2,7 @@
 import { getCollection } from 'astro:content';
 import { getDocExamplePages } from '../../../../utils/pages';
 import type { InternalFramework } from '../../../../types/ag-grid';
-import FrameworkTemplateExampleRunner from '../../../../features/example-runner/components/FrameworkTemplateExampleRunner.astro';
+import DocsFrameworkTemplate from '../../../../components/docs/DocsFrameworkTemplate.astro';
 
 export async function getStaticPaths() {
     const pages = await getCollection('docs');
@@ -14,7 +14,7 @@ export async function getStaticPaths() {
 const { internalFramework, pageName, exampleName } = Astro.params;
 ---
 
-<FrameworkTemplateExampleRunner
+<DocsFrameworkTemplate
     internalFramework={internalFramework as InternalFramework}
     pageName={pageName!}
     exampleName={exampleName!}
