+++ packages/ag-charts-website/src/pages/[internalFramework]/[pageName]/examples/[exampleName]/relative-path.astro
@@ -1,8 +1,8 @@
 ---
 import { getCollection } from 'astro:content';
 import { getDocExamplePages } from '../../../../../utils/pages';
-import FrameworkTemplateExampleRunner from '../../../../../features/example-runner/components/FrameworkTemplateExampleRunner.astro';
-import type { InternalFramework } from 'packages/ag-charts-website/src/types/ag-grid';
+import type { InternalFramework } from '../../../../../types/ag-grid';
+import DocsFrameworkTemplate from '../../../../../components/docs/DocsFrameworkTemplate.astro';
 
 /**
  * This page is the same as `[exampleName].astro`, but using relative paths for script files
@@ -22,7 +22,7 @@ export async function getStaticPaths() {
 const { internalFramework, pageName, exampleName } = Astro.params;
 ---
 
-<FrameworkTemplateExampleRunner
+<DocsFrameworkTemplate
     internalFramework={internalFramework as InternalFramework}
     pageName={pageName!}
     exampleName={exampleName!}
